ecommerce-analyst/requirements.txt
langchain>=0.3.7,<0.4.0
langchain-google-genai>=2.1.0,<3.0.0
langchain-chroma>=0.1.4,<0.2.0
langchain-community>=0.3.7,<0.4.0
chromadb>=0.5.18,<0.6.0
google-genai>=1.0.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
ragas>=0.2.5,<0.3.0
datasets>=3.1.0,<4.0.0
tiktoken>=0.8.0
sentence-transformers>=3.3.1,<4.0.0
diskcache==5.6.3
tenacity==9.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
rich>=13.9.4
pydantic>=2.10.1,<3.0.0